requirements.txt
asttokens==3.0.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
facenet-pytorch==2.6.0
filelock==3.13.1
fonttools==4.60.1
fsspec==2024.6.1
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.10
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.6.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.4
jupyter-client==8.6.3
jupyter-core==5.8.1
kiwisolver==1.4.9
lpips==0.1.4
markupsafe==2.1.5
matplotlib==3.10.6
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
ninja==1.13.0
numpy==1.26.4
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.4.0
pluggy==1.6.0
prompt-toolkit==3.0.52
psutil==7.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scipy==1.16.2
setuptools==70.2.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.3
tokenizers==0.22.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.2
triton==3.4.0
typing-extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
